Add files via upload
This folder consists of using OpenCV built in recognizer and training it on the faces dataset. Some of the results were quite weird showing incorrect labels to the faces, stating that openCV built in recognizer is not a perfect recognizer also the dataset was small, which can also be one of the reasons of the weird results.